Revert "So 91 adding to lists fixes"
