--- lib/cubit/fav_games_state.dart
@@ -1,5 +1,0 @@
-part of 'fav_games_cubit.dart';
-
-class FavGamesState extends ListState {
-  FavGamesState(List<Game> list) : super(list);
-}
