--- lib/cubit/played_games_state.dart
@@ -1,5 +1,0 @@
-part of 'played_games_cubit.dart';
-
-class PlayedGamesState extends ListState {
-  PlayedGamesState(List<Game> list) : super(list);
-}
