--- lib/cubit/wishlist_games_state.dart
@@ -1,5 +1,0 @@
-part of 'wishlist_games_cubit.dart';
-
-class WishlistGamesState extends ListState {
-  WishlistGamesState(List<Game> list) : super(list);
-}
