+++ lib/game/cover.dart
