--- lib/locale_string.dart
@@ -1,75 +1,0 @@
-import 'package:get/get.dart';
-
-class LocaleString extends Translations {
-  @override
-  Map<String, Map<String, String>> get keys => {
-        'en_US': {
-          'home': 'Home',
-          'search': 'Search',
-          'library': "Library",
-          'summary': "Summary",
-          'release_date': 'release date',
-          'genres': 'genres',
-          'game_modes': 'game modes',
-          'part_of': 'part of the series',
-          'companies': 'companies',
-          'available_on': 'available on',
-          'screenshots': 'in game screenshots',
-          'main_game': 'part of the main game',
-          'franchise': 'from the franchise',
-          'search_for': "Search for your favourite games",
-          'cancel': "Cancel",
-          'settings': "Settings",
-          'fav': "Favourite",
-          'played': "Played",
-          'wishlist': "Wishlist",
-          'search_more': "Display more results",
-          'add_fav': 'Add to favourites',
-          'add_played': 'Add to played games',
-          'add_wishlist': 'Add to wishlist',
-          'remove_fav': 'remove from favourites',
-          'remove_played': 'remove from played games',
-          'remove_wishlist': 'remove from wishlist',
-          'wishlist_removed': 'Removed Game from the wishList',
-          'wishlist_added': 'Added Game to the wishList',
-          'played_removed': 'Removed Game from the \'played\' library',
-          'played_added': 'Added Game to the \'played\' library',
-          'fav_removed': 'Removed Game from the favourites',
-          'fav_added': 'Added Game to the favourites',
-        },
-        'pl_PL': {
-          'home': 'Dla Ciebie',
-          'search': 'Szukaj',
-          'library': "Biblioteka",
-          'summary': "Opis",
-          'release_date': 'data wydania',
-          'genres': 'gatunek',
-          'game_modes': 'tryb gry',
-          'part_of': 'część serii',
-          'companies': 'firmy',
-          'available_on': 'dostępne na',
-          'screenshots': 'zrzuty ekranu z gry',
-          'main_game': 'część gry',
-          'franchise': 'pochodzi z marki',
-          'search_for': "Znajdź swoje ulubione gry",
-          'cancel': "Anuluj",
-          'settings': "Ustawienia",
-          'fav': "Ulubione",
-          'played': "Zagrane",
-          'wishlist': "Do zagrania",
-          'search_more': "Wyświetl więcej wyników",
-          'add_fav': 'Dodaj do ulubionych',
-          'add_played': 'Dodaj do zagranych',
-          'add_wishlist': 'Dodaj do \'do zagrania\'',
-          'remove_fav': 'Usuń z ulubionych',
-          'remove_played': 'Usuń z zagranych',
-          'remove_wishlist': 'Usuń z \'do zagrania\'',
-          'wishlist_removed': 'Gra została usunięta z listy do zagrania',
-          'wishlist_added': 'Gra została dodania do listy do zagrania ',
-          'played_removed': 'Gra została usinięta z \'Zagranych\'',
-          'played_added': 'Gra została dodana do \'Zagranych\'',
-          'fav_added': 'Gra została dodana do Ulubionych',
-          'fav_removed': 'Gra została usunięta z Ulubionych',
-        }
-      };
-}
