+++ lib/results_grid/results_grid.dart
@@ -1,9 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_staggered_grid_view/flutter_staggered_grid_view.dart';
-
-import 'package:game_lib_app/models/game/game.dart';
-import 'package:game_lib_app/repositories/igdb_repository.dart';
-
+import 'package:game_lib_app/resource_manager.dart';
 import 'package:lazy_load_scrollview/lazy_load_scrollview.dart';
 
 import 'results_grid_game_tile.dart';
@@ -20,9 +17,9 @@ class ResultsGrid extends StatefulWidget {
 }
 
 class _ResultsGridState extends State<ResultsGrid> {
+  ResourceManager resMan = ResourceManager();
   bool gamesLoading = false;
   int count = 0;
-  List<Game> loadedGames = [];
   Future<void> loadMoreGames() async {
     if (mounted) {
       setState(() {
@@ -30,9 +27,8 @@ class _ResultsGridState extends State<ResultsGrid> {
       });
     }
 
-    loadedGames +=
-        await IgdbRepository.fetchGamePosters(widget.whereFilters, count);
-    count = loadedGames.length;
+    await resMan.loadMoreGames(count, widget.whereFilters);
+    count = resMan.resaultsGamesLoaded.length;
     if (mounted) {
       setState(() {
         gamesLoading = false;
@@ -65,7 +61,7 @@ class _ResultsGridState extends State<ResultsGrid> {
               }
               return ResultsGameTile(
                 index: index,
-                game: loadedGames[index],
+                resourceManager: resMan,
               );
             }),
       ),
