+++ pubspec.yaml
@@ -41,10 +41,6 @@ dependencies:
   lazy_load_scrollview: ^1.3.0
   intl: ^0.17.0
   flutter_staggered_grid_view: ^0.6.2
-  dio: ^4.0.6
-  get: ^4.6.5
-  bloc: ^8.0.3
-  flutter_bloc: ^8.0.1
 
 
 dev_dependencies:
